# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Telegram Earning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register or get user from Telegram WebApp data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "200":
          description: User registered/retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get current user profile
      parameters:
        - name: telegramId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{telegramId}/daily-bonus:
    post:
      operationId: claimDailyBonus
      tags: [users]
      summary: Claim daily bonus
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Daily bonus claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyBonusResult"
  /users/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get dashboard stats for a user
      parameters:
        - name: telegramId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"

  # ========== TASKS ==========
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all active tasks for a user
      parameters:
        - name: telegramId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (admin)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task (admin)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /tasks/{taskId}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark task as complete and claim reward
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCompletion"
      responses:
        "200":
          description: Task completed, reward added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCompletionResult"

  # ========== WALLET ==========
...
  /referrals/{telegramId}:
    get:
      operationId: getReferralInfo
      tags: [referrals]
      summary: Get referral info and history
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Referral info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralInfo"

  # ========== LEADERBOARD ==========
...
  /notifications/{telegramId}:
    get:
      operationId: getUserNotifications
      tags: [notifications]
      summary: Get user notifications
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{telegramId}/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ========== SETTINGS ==========
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    User:
      type: object
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: ["string", "null"]
        coins:
          type: integer
        totalEarnings:
          type: integer
        referralCode:
          type: string
        referredBy:
          type: ["string", "null"]
        rank:
          type: string
        isBanned:
          type: boolean
        dailyBonusLastClaimed:
          type: ["string", "null"]
        streakDays:
          type: integer
        tasksCompleted:
          type: integer
        createdAt:
          type: string
        lastActive:
          type: string
      required: [id, telegramId, firstName, coins, totalEarnings, referralCode, rank, isBanned, streakDays, tasksCompleted, createdAt, lastActive]
    UserRegistration:
      type: object
      properties:
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: ["string", "null"]
        referralCode:
          type: ["string", "null"]
        initData:
          type: string
      required: [telegramId, firstName]
    UserStats:
      type: object
      properties:
        coins:
          type: integer
        totalEarnings:
          type: integer
        tasksCompleted:
          type: integer
        referralCount:
          type: integer
        rank:
          type: string
        streakDays:
          type: integer
        pendingWithdrawals:
          type: integer
        dailyBonusAvailable:
          type: boolean
        nextDailyBonusIn:
          type: ["integer", "null"]
      required: [coins, totalEarnings, tasksCompleted, referralCount, rank, streakDays, pendingWithdrawals, dailyBonusAvailable]
    DailyBonusResult:
      type: object
      properties:
        success:
          type: boolean
        coinsEarned:
          type: integer
        streakDays:
          type: integer
        totalCoins:
          type: integer
        message:
          type: string
      required: [success, coinsEarned, streakDays, totalCoins, message]
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
        reward:
          type: integer
        url:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        isActive:
          type: boolean
        isCompleted:
          type: boolean
        totalCompletions:
          type: integer
        createdAt:
          type: string
      required: [id, title, description, type, reward, isActive, isCompleted, totalCompletions, createdAt]
    TaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
        reward:
          type: integer
        url:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        isActive:
          type: boolean
      required: [title, description, type, reward, isActive]
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
        reward:
          type: integer
        url:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        isActive:
          type: boolean
    TaskCompletion:
      type: object
      properties:
        telegramId:
          type: string
        proof:
          type: ["string", "null"]
      required: [telegramId]
    TaskCompletionResult:
      type: object
      properties:
        success:
          type: boolean
        coinsEarned:
          type: integer
        totalCoins:
          type: integer
        message:
          type: string
      required: [success, coinsEarned, totalCoins, message]
...
    ReferralInfo:
      type: object
      properties:
        referralCode:
          type: string
        referralLink:
          type: string
        totalReferrals:
          type: integer
        totalEarnedFromReferrals:
          type: integer
        referrals:
          type: array
          items:
            $ref: "#/components/schemas/ReferralEntry"
      required: [referralCode, referralLink, totalReferrals, totalEarnedFromReferrals, referrals]
    ReferralEntry:
      type: object
      properties:
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: string
        joinedAt:
          type: string
        coinsEarned:
          type: integer
      required: [telegramId, firstName, joinedAt, coinsEarned]
...
    Notification:
      type: object
      properties:
        id:
          type: string
        telegramId:
          type: string
        title:
          type: string
        message:
          type: string
        type:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, telegramId, title, message, type, isRead, createdAt]
...