# lib/api-zod/src/generated/types/dailyBonusResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Earning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyBonusResult {
  success: boolean;
  coinsEarned: number;
  streakDays: number;
  totalCoins: number;
  message: string;
}