# lib/api-zod/src/generated/types/taskCompletion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Earning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskCompletion {
  telegramId: string;
  /** @nullable */
  proof?: string | null;
}